Update README.md
Successfully implemented Caesar Cipher encryption and decryption in Python!
I'm excited to share that I've been working on a Python program to encrypt and decrypt messages using the Caesar Cipher. This classic cipher technique involves shifting each letter in a message by a certain number of positions (the key) to create a scrambled message.

My Python code effectively encrypts and decrypts messages based on the provided key. I've ensured the code is efficient and handles both upper and lowercase letters.
This project has been a fantastic exercise in applying Python's capabilities for string manipulation and cryptography.

Thankful to Prodigy InfoTech for this opportunity !🎓